backend/pyproject.toml
[tool.poetry]
name = "SecObserve"
version = "1.28.1"
description = "SecObserve is an open source vulnerability management system for software development and cloud environments."
license = "BSD-3-Clause"
authors = [
    "Stefan Fleckenstein <stefan.fleckenstein@maibornwolff.de>",
]

[tool.poetry.dependencies]
python = ">= 3.10, < 3.13"
# Django
# ------------------------------------------------------------------------------
gunicorn = "23.0.0"  # https://github.com/benoitc/gunicorn
django = "5.1.6"  # https://www.djangoproject.com/
django-environ = "0.12.0"  # https://github.com/joke2k/django-environ
django-filter = "25.1"  # https://github.com/carltongibson/django-filter 
django-csp = "3.8"  # https://github.com/mozilla/django-csp
django-picklefield = "3.2"  # https://github.com/gintas/django-picklefield 
django-encrypted-model-fields = "0.6.5"  # https://gitlab.com/lansharkconsulting/django/django-encrypted-model-fields
argon2-cffi = "23.1.0"  # https://github.com/hynek/argon2_cffi
whitenoise = "6.9.0"  # https://github.com/evansd/whitenoise
# Django REST Framework
# ------------------------------------------------------------------------------
djangorestframework = "3.15.2"  # https://github.com/encode/django-rest-framework
django-cors-headers = "4.7.0" # https://github.com/adamchainz/django-cors-headers
# OpenAPI 3
# ------------------------------------------------------------------------------
drf-spectacular = "0.28.0" # https://github.com/tfranzel/drf-spectacular
drf-spectacular-sidecar = "2025.2.1"  # https://github.com/tfranzel/drf-spectacular-sidecar
# Token authentication
# ------------------------------------------------------------------------------
PyJWT = "2.10.1"  # https://github.com/jpadilla/pyjwt
# HTTP requests
# ------------------------------------------------------------------------------
requests = "2.32.3"  # https://github.com/psf/requests
# Database
# ------------------------------------------------------------------------------
pymysql = "1.1.1"  # https://github.com/PyMySQL/PyMySQL
psycopg = { version = "3.2.4", extras = ["binary"] }  # https://github.com/psycopg/psycopg
# Dataformats
# ------------------------------------------------------------------------------
defusedcsv = "2.0.0"  # https://github.com/raphaelm/defusedcsv
openpyxl = "3.1.5"  # https://foss.heptapod.net/openpyxl/openpyxl
jsonpickle = "4.0.2"  # https://github.com/jsonpickle/jsonpickle
pyyaml = "6.0.2"  # https://github.com/yaml/pyyaml
# Field specifications
# ------------------------------------------------------------------------------
packageurl-python = "0.16.0"  # https://github.com/package-url/packageurl-python
cvss = "3.4"  # https://github.com/RedHatProductSecurity/cvss
semver = "3.0.4"  # https://github.com/python-semver/python-semver
license-expression = "30.4.1"  # https://github.com/aboutcode-org/license-expression
# Task queue
# ------------------------------------------------------------------------------
huey = "2.5.2"  # https://github.com/coleifer/huey
# Issue tracker
# ------------------------------------------------------------------------------
jira = "3.8.0"  # https://github.com/pycontribs/jira
# Others
# ------------------------------------------------------------------------------
inflect = "7.5.0"  # https://github.com/jaraco/inflect
validators = "0.34.0"  # https://github.com/python-validators/validators
py-ocsf-models = "0.3.1"  # https://github.com/prowler-cloud/py-ocsf-models
spdx-tools = "0.8.3"  # https://github.com/spdx/tools-python

# Development dependencies
# ------------------------------------------------------------------------------
# Werkzeug = { version = "3.0.4", extras = ["watchdog"], optional = true }  # https://github.com/pallets/werkzeug
Werkzeug = { version = "3.1.3", optional = true }  # https://github.com/pallets/werkzeug
ipdb = { version = "0.13.13", optional = true }  # https://github.com/gotcha/ipdb
# watchgod = { version = "0.8.2", optional = true }  # https://github.com/samuelcolvin/watchgod
django-extensions = { version = "3.2.3", optional = true }  # https://github.com/django-extensions/django-extensions

# Unittest dependencies
# ------------------------------------------------------------------------------
coverage = { version = "7.6.12", optional = true }  # https://github.com/nedbat/coveragepy
django-coverage-plugin = { version = "3.1.0", optional = true }  # https://github.com/nedbat/django_coverage_plugin

# Code quality dependencies
# ------------------------------------------------------------------------------
flake8 = { version = "7.1.2", optional = true }  # https://github.com/PyCQA/flake8
flake8-isort = { version = "6.1.2", optional = true }  # https://github.com/gforcada/flake8-isort
black = { version = "25.1.0", optional = true }  # https://github.com/psf/black
pylint = { version = "3.3.4", optional = true }  # https://github.com/pylint-dev/pylint
pylint-django = { version = "2.6.1", optional = true }  # https://github.com/PyCQA/pylint-django
mypy = { version = "1.15.0", optional = true }  # https://github.com/python/mypy
django-stubs = { version = "5.1.3", optional = true }  # https://github.com/typeddjango/django-stubs
djangorestframework-stubs = { version = "3.15.2", optional = true }  # https://github.com/typeddjango/djangorestframework-stubs
types-PyMySQL = { version = "1.1.0.20241103", optional = true }  # https://github.com/python/typeshed
# import-linter = "2.0"  # https://github.com/seddonym/import-linter --> Cannot be build with Docker

[tool.poetry.extras]
dev = ["Werkzeug", "ipdb", "watchgod", "django-extensions"]
unittests = ["coverage", "django-extensions", "django-coverage-plugin"]
code_quality = ["flake8", "flake8-isort", "black", "pylint", "pylint-django", "pre-commit", "mypy", "django-stubs", "djangorestframework-stubs", "types-PyMySQL"]

[tool.black]
line-length = 120
target-version = ['py312']
